OPIK-1375 pr comments
